Fix PR3948: load_fileobj does not support reading from a pipe unless gz=False
The actual cause of the problem reported in PR3948 is that the .tell()
method of the Python sys.stdin object raises IOError.  Some random
testing suggests that all pipes raise IOError in their .tell() methods
making it currently impossible for any glue.ligolw program to read from
a pipe.  This patch fixes it by allowing the MD5File class to keep a
backup count of bytes to fall back on.  A new unit test is added to
catch this regression in the future.